+++ R/get_retrosheet.R
@@ -37,29 +37,33 @@
 #' get_retrosheet("play", 2012, "SFN")
 #' }
 #'
+#' @importFrom utils type.convert
 #' @export
 
 get_retrosheet <- function(...) {
 
+    # Warn if the function is called with stringsAsFactors = TRUE
     dots <- list(...)
-
-    if (!is.null(dots[["stringsAsFactors"]])) {
-        if (dots[["stringsAsFactors"]]) {
-            warning("get_retrosheet() forces stringsAsFactors = FALSE. For stringsAsFactors = TRUE, use getRetrosheet() instead")
-        }
+    if (isTRUE(dots[["stringsAsFactors"]])) {
+        warning("get_retrosheet() yields unexpected results with stringsAsFactors = TRUE. For more predictable behavior, use getRetroshet()")
+        stringsAsFactors <- TRUE
+    } else {
+        stringsAsFactors <- FALSE
     }
 
+    # Call the original function with all of the same arguments
     response <- getRetrosheet(...)
 
     matrix_to_df <- function(x) {
 
         # If the response is a single matrix, convert it to a tibble
-        if (is.matrix(x)) x <- as.data.frame(x, stringsAsFactors = FALSE)
+        if (is.matrix(x)) x <- as.data.frame(x, stringsAsFactors = stringsAsFactors)
 
         # If the response is a dataframe (or was a matrix and has been converted to a dataframe)
         if (is.data.frame(x)) {
-            out <- lapply(x, type.convert, na.strings = c("", "NA"), as.is = TRUE)
 
+            # Aggressively guess the data type via type.convert
+            out <- lapply(x, type.convert, na.strings = c("", "NA"), as.is = TRUE)
 
             # If there's a date column, make sure it's Date-type
             if ("Date" %in% names(out)) {
@@ -67,12 +71,11 @@ get_retrosheet <- function(...) {
                 out$Date <- as.Date(out$Date, tryFormats = c("%Y-%m-%d", "%Y%m%d"))
             }
 
-            out <- as.data.frame(out, stringsAsFactors = FALSE)
-            return(out)
+            return(as.data.frame(out, stringsAsFactors = stringsAsFactors))
 
         # If the response is a list of objects, re-apply this function to all of the objects in the list
         } else if (length(x) > 1) {
-           lapply(x, matrix_to_df)
+            lapply(x, matrix_to_df)
         } else {
             return(x) # If the response is not a list, matrix, or df, return the input
         }