+++ NEWS.md
@@ -2,6 +2,7 @@
   * Replaced kludge-y `closeAllConnections()` with `close()` calls to specific connections ([#1](https://github.com/colindouglas/retrosheet/issues/1))
   * Added some rudimentary data cleanup to the `get_retrosheet()` wrapper function
     * Dates in retrosheet data now return date-type data, everything else is converted from `chr` via the relatively aggressive `type.convert()`
+  * Removed `data.table` dependency, replaced `data.table::fread()` with `base::read.csv()` calls
     
 # retrosheet 1.1.1
   * Added `get_retrosheet()`, a wrapper for `getRetrosheet()` that takes the same arguments and returns the same data, except coercing matrices into tibbles.